+++ x/ethbridge/module.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
+	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/supply"
 )
 
@@ -74,23 +75,25 @@ type AppModule struct {
 	AppModuleBasic
 	AppModuleSimulation
 
-	OracleKeeper oracle.Keeper
-	SupplyKeeper supply.Keeper
-	Codespace    sdk.CodespaceType
-	Codec        *codec.Codec
+	OracleKeeper  oracle.Keeper
+	SupplyKeeper  supply.Keeper
+	AccountKeeper auth.AccountKeeper
+	Codespace     sdk.CodespaceType
+	Codec         *codec.Codec
 }
 
 // NewAppModule creates a new AppModule object
-func NewAppModule(oracleKeeper oracle.Keeper, supplyKeeper supply.Keeper, codespace sdk.CodespaceType, cdc *codec.Codec) AppModule {
+func NewAppModule(oracleKeeper oracle.Keeper, supplyKeeper supply.Keeper, accountKeeper auth.AccountKeeper, codespace sdk.CodespaceType, cdc *codec.Codec) AppModule {
 
 	return AppModule{
 		AppModuleBasic:      AppModuleBasic{},
 		AppModuleSimulation: AppModuleSimulation{},
 
-		OracleKeeper: oracleKeeper,
-		SupplyKeeper: supplyKeeper,
-		Codespace:    codespace,
-		Codec:        cdc,
+		OracleKeeper:  oracleKeeper,
+		SupplyKeeper:  supplyKeeper,
+		AccountKeeper: accountKeeper,
+		Codespace:     codespace,
+		Codec:         cdc,
 	}
 }
 
@@ -110,7 +113,7 @@ func (AppModule) Route() string {
 
 // NewHandler returns an sdk.Handler for the ethbridge module.
 func (am AppModule) NewHandler() sdk.Handler {
-	return NewHandler(am.OracleKeeper, am.SupplyKeeper, am.Codespace, am.Codec)
+	return NewHandler(am.OracleKeeper, am.SupplyKeeper, am.AccountKeeper, am.Codespace, am.Codec)
 }
 
 // QuerierRoute returns the ethbridge module's querier route name.